requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
